requirements.txt
praw
python-dotenv
colorama